# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "go track",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://accounts.google.com/o/oauth2/v2/auth",
            "tokenUrl": "https://oauth2.googleapis.com/token",
            "scopes": {
              "openid": "OpenID Connect scope",
              "email": "Access to your email address",
              "profile": "Access to your basic profile info"
            }
          }
        }
      }
    },
    "schemas": {}
  },
  "paths": {
    "/sessions": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Authenticate user google",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tokenId": {
                    "type": "string"
                  }
                },
                "required": [
                  "tokenId"
                ],
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string",
                      "pattern": "^[A-Za-z0-9-_]+\\.[A-Za-z0-9-_]+\\.[A-Za-z0-9-_]*$"
                    }
                  },
                  "required": [
                    "accessToken"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/expenses": {
      "post": {
        "tags": [
          "Expense"
        ],
        "description": "create expense",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 4
                  },
                  "value": {
                    "type": "number",
                    "exclusiveMinimum": true,
                    "minimum": 1
                  },
                  "service": {
                    "type": "string",
                    "minLength": 2
                  }
                },
                "required": [
                  "title",
                  "value",
                  "service"
                ],
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "expenseId": {
                      "type": "string",
                      "pattern": "^[0-9a-z]+$"
                    }
                  },
                  "required": [
                    "expenseId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "default": "\tUnauthorized"
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Expense"
        ],
        "description": "get list expenses by month",
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "month",
            "required": true
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "year",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "expenses": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "userId": {
                            "type": "string",
                            "pattern": "^[cC][^\\s-]{8,}$"
                          },
                          "id": {
                            "type": "string",
                            "pattern": "^[cC][^\\s-]{8,}$"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "title": {
                            "type": "string"
                          },
                          "value": {
                            "type": "number",
                            "exclusiveMinimum": true,
                            "minimum": 0
                          },
                          "service": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "userId",
                          "id",
                          "createdAt",
                          "title",
                          "value",
                          "service"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "expenses"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "default": "\tUnauthorized"
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  }
}